Get projects data via api and update front-end to ingest
added new dependencies markdown, graphql, and graphql-request

Added margin to layout header, removed margin from layout child components, made scroll to top a component and nested in layout, cleaned up functions

Removed /data and files all data now handled through graphCMS or inline within component
