+++ package.json
@@ -9,8 +9,9 @@
     "@react-pdf-viewer/core": "^2.4.3",
     "@react-pdf-viewer/default-layout": "^2.4.3",
     "axios": "^0.21.1",
-    "dns-packet": "^5.2.2",
-    "multicast-dns": "^7.2.2",
+    "graphql": "^15.5.1",
+    "graphql-request": "^3.4.0",
+    "markdown-to-jsx": "^7.1.3",
     "pdfjs-dist": "^2.6.347",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
@@ -39,5 +40,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "devDependencies": {}
 }
