+++ src/App.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import { BrowserRouter as Router, } from "react-router-dom";
+import { BrowserRouter as Router } from "react-router-dom";
 import AppContext from './providers/AppContext';
 import Routes from './Routes';
 
 export default function App() {
+
     return (
       <Router forceRefresh={true}>
         <AppContext>
