+++ src/Routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Switch, Route, Redirect} from 'react-router-dom';
 
-import { About, Contact, Home, NotFound, Projects, Resume } from './pages';
+import { About, Contact, Home, NotFound, Project, Resume } from './pages';
 
 export default function Routes(props) {
 
@@ -13,7 +13,7 @@ export default function Routes(props) {
             <Route path="/contact" exact render={(props) => <Contact {...props} title="Contact" /> } />
             <Route path="/resume" exact render={(props) => <Resume {...props} title="Resume" /> } />
             <Route path="/not-found" exact render={(props) => <NotFound {...props} title="404" /> } />
-            <Route path="/:link" render={(props) => <Projects {...props} title="Projects" /> } />
+            <Route path="/:slug" render={(props) => <Project {...props} title="Project" /> } />
             <Redirect from="*" to="/not-found" />
         </Switch>
     )
