+++ src/components/ContactForm.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { apiPost } from '../functions/api';
-import { Data as ContactFields } from '../data/ContactForm';
 
 import { Typography, Grid, TextField, Button } from '@material-ui/core';
 import Alert from './Alert';
@@ -8,9 +7,7 @@ import Alert from './Alert';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles(theme => ({
-    root: {
-        marginTop: '72px',
-    },
+    root: {},
     form: {
         marginTop: '32px',
     },
@@ -46,6 +43,51 @@ const initialState = {
 export default function ContactForm() {
     const classes = useStyles();
     const [state, setState] = React.useState(initialState);
+
+    const ContactFields = [
+        {
+            gridWidth: 6,
+            type: "text",
+            name: "first name",
+            id: "firstName",
+            variant: "outlined",
+            label: "First Name",
+            required: true,
+        },
+        {
+            gridWidth: 6,
+            type: "text",
+            name: "last name",
+            id: "lastName",
+            variant: "outlined",
+            label: "Last Name",
+        },
+        {
+            type: "text",
+            name: "company",
+            id: "company",
+            variant: "outlined",
+            label: "Company",
+        },
+        {
+            type: "email",
+            name: "email",
+            id: "email",
+            variant: "outlined",
+            label: "Email",
+            required: true,
+        },
+        {
+            type: "text",
+            name: "message",
+            id: "message",
+            variant: "outlined",
+            label: "Details about your company, project, or team",
+            multiline: true,
+            rows: 8,
+            required: true,
+        }
+    ];
 
     const handleChange = e => {
         setState({ ...state, [e.target.id]: e.target.value });
