+++ src/components/Layout.js
@@ -2,13 +2,15 @@ import React from 'react';
 
 import Container from '@material-ui/core/Container';
 
-import { Header, Footer } from './';
+import { Header, ScrollToTop, Footer } from './';
 
 export default function Layout(props) {
+    Layout.defaultProps = { scroll: true };
     return (
         <Container>
             <Header />
             {props.children}
+            {props.scroll && <ScrollToTop />}
             <Footer />
         </Container>
     );
