--- src/data/ContactForm.js
@@ -1,44 +1,0 @@
-export const Data = [
-    {
-        gridWidth: 6,
-        type: "text",
-        name: "first name",
-        id: "firstName",
-        variant: "outlined",
-        label: "First Name",
-        required: true,
-    },
-    {
-        gridWidth: 6,
-        type: "text",
-        name: "last name",
-        id: "lastName",
-        variant: "outlined",
-        label: "Last Name",
-    },
-    {
-        type: "text",
-        name: "company",
-        id: "company",
-        variant: "outlined",
-        label: "Company",
-    },
-    {
-        type: "email",
-        name: "email",
-        id: "email",
-        variant: "outlined",
-        label: "Email",
-        required: true,
-    },
-    {
-        type: "text",
-        name: "message",
-        id: "message",
-        variant: "outlined",
-        label: "Details about your company, project, or team",
-        multiline: true,
-        rows: 8,
-        required: true,
-    }
-];
