--- src/data/index.js
@@ -1,5 +1,0 @@
-import { ProjectsData } from './ProjectsData'
-
-export {
-    ProjectsData
-}
